pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "robot_comic"
version = "0.6.2"
authors = [{ name = "Pollen Robotics", email = "contact@pollen-robotics.com" }]
description = "Robot Comic voice app for Reachy Mini"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    #Media
    "av",
    "aiortc>=1.13.0",
    "fastrtc>=0.0.34",
    "gradio==5.50.1.dev1",
    "huggingface-hub==1.3.0",
    "httpx",
    "mediapipe==0.10.14",

    #Environment variables
    "python-dotenv",

    #OpenAI
    "openai>=2.30.0",

    #Gemini
    "google-genai>=1.0.0",

    #OpenTelemetry (always installed; gated by ROBOT_INSTRUMENTATION at runtime)
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",

    #Reachy mini
    "reachy_mini_dances_library",
    "reachy_mini_toolbox",
    "reachy-mini>=1.7.1",
    "gradio_client>=1.13.3",
]

[project.optional-dependencies]
local_vision = [
  "accelerate>=1.13.0",
  "torch>=2.1",
  "transformers==5.3.0",
  "num2words",
]
yolo_vision = [
  "ultralytics",
  "supervision",
  "opencv-python>=4.13.0.92",
]
mediapipe_vision = [
  # Kept for backwards-compatible install commands. MediaPipe is now part of
  # the default app install because it powers the lightweight head tracker.
]
local_stt = [
  "moonshine-voice>=0.0.46",
]
otel = [
  # OTLP exporter for remote/SigNoz export (ROBOT_INSTRUMENTATION=remote).
  # Requires protobuf>=5 which conflicts with mediapipe<5; install separately
  # on hosts that don't use mediapipe (e.g. CI, cloud aggregators).
  "opentelemetry-exporter-otlp-proto-grpc>=1.24.0",
]
monitor = [
  "rich>=13.0",
]
all_vision = [
  "accelerate>=1.13.0",
  "torch>=2.1",
  "transformers==5.3.0",
  "num2words",
  "ultralytics",
  "supervision",
  "opencv-python>=4.13.0.92",
]

[dependency-groups]
dev = [
  "pytest",
  "pytest-asyncio",
  "ruff==0.12.0",
  "mypy==1.18.2",
  "pre-commit",
  "types-requests",
  "python-semantic-release>=10.5.3",
]

[project.scripts]
robot-comic = "robot_comic.main:main"
robot-comic-monitor = "robot_comic.monitor:main"

[tool.uv]
torch-backend = "cu124"
dependency-metadata = [
    # Upstream metadata currently marks `gstreamer-msvc-runtime` as unconditional.
    # It should only be required on Windows. Temp fix.
    { name = "gstreamer-libs", version = "1.28.1", requires-dist = ["gstreamer-msvc-runtime; sys_platform == 'win32'", "setuptools"] },
]

[project.entry-points."reachy_mini_apps"]
robot_comic = "robot_comic.main:RobotComic"

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
robot_comic = [
  "images/*",
  "static/*",
  ".env.example",
  "demos/**/*.txt",
  "prompts_library/*.txt",
  "prompts/**/*.txt",
]
reachy_talk_data = [
  "profiles/**/*.txt",
  "profiles/**/*.py",
]

[tool.ruff]
line-length = 119
exclude = [".venv", "dist", "build", "**/__pycache__", "*.egg-info", ".mypy_cache", ".pytest_cache"]

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "F",   # pyflakes
  "W",   # pycodestyle warnings
  "I",   # isort
  "C4",  # flake8-comprehensions
  "D",   # pydocstyle
]
ignore = [
  "E501",  # handled by formatter
  "D100",  # ignore missing module docstrings
  "D203",  # blank line before class docstring (conflicts with D211)
  "D213",  # summary on second line (conflicts with D212)
]

[tool.ruff.lint.isort]
length-sort = true
lines-after-imports = 2
no-lines-before = ["standard-library", "local-folder"]
known-local-folder = ["robot_comic"]
known-first-party = ["reachy_mini", "reachy_mini_dances_library", "reachy_mini_toolbox"]
split-on-trailing-comma = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
files = ["src/"]
ignore_missing_imports = true
strict = true
show_error_codes = true
warn_unused_ignores = true
